compiler: faster string-heavy codegen + two correctness fixes
Correctness: type string/array .length as i32 (was an unusable i64, breaking
i<x.length index loops), and decode string escapes (\n \t \r \0 \ " \') at
codegen (were emitted literally). Performance: route runtime string allocations
through a shared arena instead of page_allocator; flatten chained concats into a
single concat; compile append-only string locals into growable ArrayList buffers;
and destination-passing for builder functions (append into a caller buffer rather
than build-and-copy). Together ~60x faster on string-builder workloads; all
conservative (fall back to prior codegen unless provably safe).

+++ src/lumen_check.zig
@@ -2650,7 +2650,10 @@ const Checker = struct {
                 }
                 if ((types.isStringLike(obj_type) or types.isArray(obj_type)) and std.mem.eql(u8, field.name, "length")) {
                     field.builtin = .length;
-                    return .i64;
+                    // `int` (i32) is the language's integer; typing length as i32
+                    // lets the common `for`/`while (i < x.length)` index idiom and
+                    // `charAt(i)`/`substring(...)` compose without an unusable i64.
+                    return .i32;
                 }
                 if ((types.isMap(obj_type) or types.isSet(obj_type)) and std.mem.eql(u8, field.name, "size")) {
                     field.builtin = .container_size;
